Experiment with kysely for queries. Generate kysely types from prisma.schema.
We are still using prisma to generate the kyselt types from
prisma.schema (using the prisma-kysely tool). This is because I haven't
migrated all the auth code from the epic stack from prisma to kysely
yet. Also I am not convinced that using Prisma for schema defs and
kylesy for queries, for the sorts of reasons documented in https://codedamn.com/news/product/dont-use-prisma