Fix hydration errors (fixes #28)
This was caused by componenets rendering 'Bad' HTML (by react's
definition). Specifically, <a> tags rendered inside <a> tags. This
happened because in Note attachments and prent threads, the entire post
is a clickable link, but the rendered markdown may also contain links.
Solution is to tell Markdown renderer not to render links (render them
as <spans> instead). Custom Markdown component with deactivateLinks
option solves this.