base nix on ubuntu, no nix-shell for prod image, prod-image 1.47GB (#100)
the production docker image doen't use the nix-shell anymore. We only
need the packages without env variables, since nothing needs to be
compiled. We also base it on ubuntu, because it is a smaller base system
(~25MB) than the nix image was (~200MB). nix garbage collection helps to
reduce the nix-store even more.

This brings the production docker-image down from **2.01GB** to
**1.47GB** (save **0.54GB**).

No more `/bin/sh` is needed in the production shell, since we're not
using a nix-shell anymore.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>